country-levels.js moved to geography
