+++ src/events/crawler/scrape-data/run-scraper.js
@@ -1,9 +1,9 @@
 import path from 'path';
 import datetime from '../../../shared/lib/datetime/index.js';
-import log from '../../../shared/lib/log.js';
-import * as geography from '../../../shared/lib/geography/index.js';
 import reporter from '../../../shared/lib/error-reporter.js';
-import * as countryLevels from '../../../shared/lib/country-levels.js';
+import * as countryLevels from '../../../shared/lib/geography/country-levels.js';
+import * as geography from '../../../shared/lib/geography/index.js';
+import log from '../../../shared/lib/log.js';
 
 const numericalValues = ['cases', 'tested', 'recovered', 'deaths', 'active'];
