+++ src/events/processor/find-features/index.js
@@ -1,12 +1,12 @@
 import geoTz from 'geo-tz';
 import { join } from 'path';
-import * as turf from '../../../shared/lib/geography/turf.js';
+import reporter from '../../../shared/lib/error-reporter.js';
 import * as fs from '../../../shared/lib/fs.js';
+import * as countryLevels from '../../../shared/lib/geography/country-levels.js';
+import * as geography from '../../../shared/lib/geography/index.js';
+import * as turf from '../../../shared/lib/geography/turf.js';
 import log from '../../../shared/lib/log.js';
 import espGeoJson from '../vendor/esp.json';
-import * as geography from '../../../shared/lib/geography/index.js';
-import reporter from '../../../shared/lib/error-reporter.js';
-import * as countryLevels from '../../../shared/lib/country-levels.js';
 
 const DEBUG = false;
