+++ src/events/processor/find-populations/index.js
@@ -1,9 +1,9 @@
 import { join, resolve } from 'path';
-import * as countryLevels from '../../../shared/lib/country-levels.js';
-import * as fs from '../../../shared/lib/fs.js';
-import log from '../../../shared/lib/log.js';
-import * as geography from '../../../shared/lib/geography/index.js';
 import reporter from '../../../shared/lib/error-reporter.js';
+import * as fs from '../../../shared/lib/fs.js';
+import * as countryLevels from '../../../shared/lib/geography/country-levels.js';
+import * as geography from '../../../shared/lib/geography/index.js';
+import log from '../../../shared/lib/log.js';
 
 const dataPath = join(__dirname, '..', 'vendor', 'population');
